Distance Calculations

A function to calculate the Euclidean distance between two vectors.

Usage

euclidean_distance(.data, .x, .y, .pull_vector = FALSE)

Arguments

.data

A data frame

.x

A numeric vector

.y

A numeric vector

.pull_vector

A boolean of TRUE or FALSE. Default is FALSE which will augment the distance to the data frame. TRUE will return a vector of the distances as the return.

Value

A numeric Vector of ditances

Details

A function to calculate the Euclidean distance between two vectors. It uses the formula sqrt((x - lag(x))^2 + (y - lag(y))^2). The function uses augments the data frame with a new column called distance.

See also

Author

Steven P. Sanderson II, MPH

Examples

set.seed(123)
df <- rw30()
euclidean_distance(df, x, y)
#> Error in dplyr::mutate(dplyr::as_tibble(.data), distance = sqrt((!!x_var -     dplyr::lag(!!x_var, 1))^2 + (!!y_var - dplyr::lag(!!y_var,     1))^2)): ℹ In argument: `distance = sqrt((x - dplyr::lag(x, 1))^2 + (y -
#>   dplyr::lag(y, 1))^2)`.
#> Caused by error:
#> ! object 'x' not found
euclidean_distance(df, x, y, TRUE) |> head(10)
#> Error in dplyr::mutate(dplyr::as_tibble(.data), distance = sqrt((!!x_var -     dplyr::lag(!!x_var, 1))^2 + (!!y_var - dplyr::lag(!!y_var,     1))^2)): ℹ In argument: `distance = sqrt((x - dplyr::lag(x, 1))^2 + (y -
#>   dplyr::lag(y, 1))^2)`.
#> Caused by error:
#> ! object 'x' not found