Subset Walks by Extreme Values

This function subsets random walks to identify the walk with the maximum or minimum value.

Usage

subset_walks(.data, .type = "max")

Arguments

.data

A data frame containing random walks. It must have columns walk_number and y.

.type

A character string specifying the type of subset: "max" for maximum value, "min" for minimum value, or "both" for both maximum and minimum values.

Value

A data frame containing the subset walk.

Details

The subset_walks function takes a data frame containing random walks and subsets it to return the walk with the maximum or minimum value based on the specified type. It requires that the input data frame contains columns walk_number and y.

See also

Author

Steven P. Sanderson II, MPH

Examples

df <- rw30()
subset_walks(df, .type = "max")
#> # A tibble: 100 × 3
#>    walk_number step_number      y
#>    <fct>             <int>  <dbl>
#>  1 15                    1  0    
#>  2 15                    2 -0.786
#>  3 15                    3  1.50 
#>  4 15                    4  0.405
#>  5 15                    5  0.619
#>  6 15                    6  1.51 
#>  7 15                    7  2.53 
#>  8 15                    8  3.62 
#>  9 15                    9  3.46 
#> 10 15                   10  2.64 
#> # ℹ 90 more rows
subset_walks(df, .type = "min")
#> # A tibble: 100 × 3
#>    walk_number step_number     y
#>    <fct>             <int> <dbl>
#>  1 21                    1  0   
#>  2 21                    2 -1.32
#>  3 21                    3 -1.46
#>  4 21                    4 -2.33
#>  5 21                    5 -3.13
#>  6 21                    6 -2.08
#>  7 21                    7 -1.91
#>  8 21                    8 -2.95
#>  9 21                    9 -3.42
#> 10 21                   10 -3.70
#> # ℹ 90 more rows
subset_walks(df, .type = "both")
#> # A tibble: 200 × 3
#>    walk_number step_number      y
#>    <fct>             <int>  <dbl>
#>  1 15                    1  0    
#>  2 15                    2 -0.786
#>  3 15                    3  1.50 
#>  4 15                    4  0.405
#>  5 15                    5  0.619
#>  6 15                    6  1.51 
#>  7 15                    7  2.53 
#>  8 15                    8  3.62 
#>  9 15                    9  3.46 
#> 10 15                   10  2.64 
#> # ℹ 190 more rows