Summarize Walks Data

Summarizes random walk data by computing statistical measures.

Usage

summarize_walks(.data, .value, .group_var)

summarise_walks(.data, .value, .group_var)

Arguments

.data

A data frame or tibble containing random walk data.

.value

A column name (unquoted) representing the value to summarize.

.group_var

A column name (unquoted) representing the grouping variable.

Value

A tibble containing the summarized statistics for each group, including mean, median, range, quantiles, variance, standard deviation, and more.

Details

This function requires that the input data frame contains a column named 'walk_number' and that the value to summarize is provided. It computes statistics such as mean, median, variance, and quantiles for the specified value variable. #' This function summarizes a data frame containing random walk data by computing various statistical measures for a specified value variable, grouped by a specified grouping variable. It checks for necessary attributes and ensures that the data frame is structured correctly.

Author

Steven P. Sanderson II, MPH

Examples

library(dplyr)
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

# Example data frame
walk_data <- random_normal_walk(.initial_value = 100)

# Summarize the walks
summarize_walks(walk_data, cum_sum, walk_number)
#> Registered S3 method overwritten by 'quantmod':
#>   method            from
#>   as.zoo.data.frame zoo 
#> # A tibble: 25 × 17
#>    walk_number fns         fns_name dimensions mean_val median range quantile_lo
#>    <fct>       <chr>       <chr>         <dbl>    <dbl>  <dbl> <dbl>       <dbl>
#>  1 1           random_nor… Random …          1    100.   100.  1.47         99.5
#>  2 2           random_nor… Random …          1     99.8   99.8 0.721        99.6
#>  3 3           random_nor… Random …          1     99.5   99.5 0.987        99.2
#>  4 4           random_nor… Random …          1     98.8   98.6 1.96         98.1
#>  5 5           random_nor… Random …          1    101.   101.  1.41        100. 
#>  6 6           random_nor… Random …          1    100.   100.  0.937        99.6
#>  7 7           random_nor… Random …          1    101.   101.  1.94        100. 
#>  8 8           random_nor… Random …          1     99.4   99.5 1.70         98.6
#>  9 9           random_nor… Random …          1    101.   100.  2.03         99.9
#> 10 10          random_nor… Random …          1    101.   100.  1.22        100. 
#> # ℹ 15 more rows
#> # ℹ 9 more variables: quantile_hi <dbl>, variance <dbl>, sd <dbl>,
#> #   min_val <dbl>, max_val <dbl>, harmonic_mean <dbl>, geometric_mean <dbl>,
#> #   skewness <dbl>, kurtosis <dbl>
summarize_walks(walk_data, y) |>
  glimpse()
#> Warning: There was 1 warning in `dplyr::summarize()`.
#> ℹ In argument: `geometric_mean = exp(mean(log(y)))`.
#> Caused by warning in `log()`:
#> ! NaNs produced
#> Rows: 1
#> Columns: 16
#> $ fns            <chr> "random_normal_walk"
#> $ fns_name       <chr> "Random Normal Walk"
#> $ dimensions     <dbl> 1
#> $ mean_val       <dbl> -0.002189823
#> $ median         <dbl> -0.005327023
#> $ range          <dbl> 0.6899619
#> $ quantile_lo    <dbl> -0.1991602
#> $ quantile_hi    <dbl> 0.1989294
#> $ variance       <dbl> 0.01035473
#> $ sd             <dbl> 0.1017582
#> $ min_val        <dbl> -0.3047861
#> $ max_val        <dbl> 0.3851758
#> $ harmonic_mean  <dbl> -0.01258915
#> $ geometric_mean <dbl> NaN
#> $ skewness       <dbl> 0.1464476
#> $ kurtosis       <dbl> 0.09097494

# Example with missing value variable
# summarize_walks(walk_data, NULL, group) # This will trigger an error.