Summarize Walks Data

Summarizes random walk data by computing statistical measures.

Usage

summarize_walks(.data, .value, .group_var)

summarise_walks(.data, .value, .group_var)

Arguments

.data

A data frame or tibble containing random walk data.

.value

A column name (unquoted) representing the value to summarize.

.group_var

A column name (unquoted) representing the grouping variable.

Value

A tibble containing the summarized statistics for each group, including mean, median, range, quantiles, variance, standard deviation, and more.

Details

This function requires that the input data frame contains a column named 'walk_number' and that the value to summarize is provided. It computes statistics such as mean, median, variance, and quantiles for the specified value variable. #' This function summarizes a data frame containing random walk data by computing various statistical measures for a specified value variable, grouped by a specified grouping variable. It checks for necessary attributes and ensures that the data frame is structured correctly.

Author

Steven P. Sanderson II, MPH

Examples

library(dplyr)
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

# Example data frame
walk_data <- random_normal_walk(.initial_value = 100)

# Summarize the walks
summarize_walks(walk_data, cum_sum_y, walk_number) |>
 glimpse()
#> Registered S3 method overwritten by 'quantmod':
#>   method            from
#>   as.zoo.data.frame zoo 
#> Rows: 25
#> Columns: 17
#> $ walk_number    <fct> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, …
#> $ fns            <chr> "random_normal_walk", "random_normal_walk", "random_nor…
#> $ fns_name       <chr> "Random Normal Walk", "Random Normal Walk", "Random Nor…
#> $ obs            <int> 100, 100, 100, 100, 100, 100, 100, 100, 100, 100, 100, …
#> $ mean_val       <dbl> 99.00097, 100.62544, 99.25515, 100.92042, 99.54205, 99.…
#> $ median         <dbl> 99.11091, 100.73591, 99.34099, 100.87174, 99.53793, 99.…
#> $ range          <dbl> 2.4465330, 1.8085902, 1.9100238, 1.5094297, 1.3299500, …
#> $ quantile_lo    <dbl> 97.75175, 99.80969, 98.20988, 100.31400, 98.92827, 99.3…
#> $ quantile_hi    <dbl> 99.96255, 101.40885, 100.01079, 101.53776, 100.05828, 1…
#> $ variance       <dbl> 0.45384343, 0.24222090, 0.31814411, 0.12433796, 0.10091…
#> $ sd             <dbl> 0.6779294, 0.4952646, 0.5676013, 0.3548406, 0.3196762, …
#> $ min_val        <dbl> 97.62691, 99.70515, 98.13139, 100.11858, 98.75010, 99.3…
#> $ max_val        <dbl> 100.07345, 101.51374, 100.04141, 101.62801, 100.08005, …
#> $ harmonic_mean  <dbl> 98.99637, 100.62303, 99.25193, 100.91919, 99.54103, 99.…
#> $ geometric_mean <dbl> 98.99867, 100.62423, 99.25354, 100.91980, 99.54154, 99.…
#> $ skewness       <dbl> -0.34582728, -0.27031231, -0.50457492, 0.08254814, -0.3…
#> $ kurtosis       <dbl> -1.16355761, -0.99910242, -0.81988631, -0.76025445, -0.…
summarize_walks(walk_data, y) |>
  glimpse()
#> Warning: There was 1 warning in `dplyr::summarize()`.
#> ℹ In argument: `geometric_mean = exp(mean(log(y)))`.
#> Caused by warning in `log()`:
#> ! NaNs produced
#> Rows: 1
#> Columns: 16
#> $ fns            <chr> "random_normal_walk"
#> $ fns_name       <chr> "Random Normal Walk"
#> $ obs            <int> 100
#> $ mean_val       <dbl> -0.004358458
#> $ median         <dbl> -0.00267075
#> $ range          <dbl> 0.6382071
#> $ quantile_lo    <dbl> -0.1973014
#> $ quantile_hi    <dbl> 0.1895718
#> $ variance       <dbl> 0.009922753
#> $ sd             <dbl> 0.09963793
#> $ min_val        <dbl> -0.3146528
#> $ max_val        <dbl> 0.3235543
#> $ harmonic_mean  <dbl> -0.1461506
#> $ geometric_mean <dbl> NaN
#> $ skewness       <dbl> -0.07339031
#> $ kurtosis       <dbl> 0.03030584

# Example with missing value variable
# summarize_walks(walk_data, NULL, group) # This will trigger an error.